move grammar into its own file
